Restructure refinement setup, improve use of parameters
